# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Echo Sample API
  version: "1.0.0"
paths:
  /health:
    get:
      summary: Health check
      description: Returns application health status.
      operationId: getHealth
      responses:
        "200":
          description: Health status payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /users:
    post:
      summary: Create user
      description: Create a new user.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      summary: Get user
      description: Retrieve a user by ID.
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Retrieved user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found.
components:
  schemas:
    Health:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: ok
    User:
      type: object
      required:
        - id
        - name
        - email
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com
    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          example: Jane Doe
        email:
          type: string
          format: email
          example: jane@example.com